Separate test for linkedlist
