+++ linkedlist/is_sorted.py
@@ -8,15 +8,6 @@ Null :List is sorted
 1 2 3 4 :List is sorted
 1 2 -1 3 :List is not sorted
 """
-import unittest
-
-
-class Node:
-    def __init__(self, x):
-        self.val = x
-        self.next = None
-
-
 def is_sorted(head):
     if not head:
         return True
@@ -26,31 +17,3 @@ def is_sorted(head):
             return False
         current = current.next
     return True
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_is_sorted(self):
-
-        head = Node(-2)
-        head.next = Node(2)
-        head.next.next = Node(2)
-        head.next.next.next = Node(4)
-        head.next.next.next.next = Node(9)
-
-        # head -> -2 -> 2 -> 2 -> 4 -> 9
-        self.assertTrue(is_sorted(head))
-
-        head = Node(1)
-        head.next = Node(2)
-        head.next.next = Node(8)
-        head.next.next.next = Node(4)
-        head.next.next.next.next = Node(6)
-
-        # head -> 1 -> 2 -> 8 -> 4 -> 6
-        self.assertFalse(is_sorted(head))
-
-
-if __name__ == "__main__":
-
-    unittest.main()
