+++ linkedlist/swap_in_pairs.py
@@ -9,37 +9,22 @@ Your algorithm should use only constant space.
 You may not modify the values in the list,
 only nodes itself can be changed.
 """
-
-
-class Node:
-    def __init__(self, x=0):
+class Node(object):
+    def __init__(self, x):
         self.val = x
         self.next = None
 
-
-def swap_pairs(head:"Node")->"Node":
+def swap_pairs(head):
     if not head:
         return head
-    start = Node()
-    pre = start
-    pre.next = head
-    while pre.next and pre.next.next:
-        a = pre.next
-        b = pre.next.next
-        pre.next, a.next, b.next = b, b.next, a
-        pre = a
+    start = Node(0)
+    start.next = head
+    current = start
+    while current.next and current.next.next:
+        first = current.next
+        second = current.next.next
+        first.next = second.next
+        current.next = second
+        current.next.next = first
+        current = current.next.next
     return start.next
-
-
-if __name__ == "__main__":
-    n = Node(1)
-    n.next = Node(2)
-    n.next.next = Node(3)
-    n.next.next.next = Node(4)
-    res = swap_pairs(n)
-
-    while res:
-        print(res.val, end=" ")
-        res = res.next
-    print("should be 2 1 4 3 ")
-
