+++ linkedlist/test_linkedlist.py
@@ -1,0 +1,124 @@
+from reverse import reverse_list, reverse_list_recursive
+from is_sorted import is_sorted
+from remove_range import remove_range
+from swap_in_pairs import swap_pairs
+from rotate_list import rotate_right
+from is_cyclic import is_cyclic
+
+import unittest
+
+class Node(object):
+    def __init__(self, x):
+        self.val = x
+        self.next = None
+
+# Convert from linked list Node to list for testing
+def convert(head):
+    ret = []
+    if head:
+        current = head
+        while current:
+            ret.append(current.val)
+            current = current.next
+    return ret
+
+class TestSuite(unittest.TestCase):
+    def test_reverse_list(self):
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(3)
+        head.next.next.next = Node(4)
+        self.assertEqual([4, 3, 2, 1], convert(reverse_list(head)))
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(3)
+        head.next.next.next = Node(4)
+        self.assertEqual([4, 3, 2, 1], convert(reverse_list_recursive(head)))
+
+    def test_is_sorted(self):
+        head = Node(-2)
+        head.next = Node(2)
+        head.next.next = Node(2)
+        head.next.next.next = Node(4)
+        head.next.next.next.next = Node(9)
+        # head -> -2 -> 2 -> 2 -> 4 -> 9
+        self.assertTrue(is_sorted(head))
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(8)
+        head.next.next.next = Node(4)
+        head.next.next.next.next = Node(6)
+        # head -> 1 -> 2 -> 8 -> 4 -> 6
+        self.assertFalse(is_sorted(head))
+
+    def test_remove_range(self):
+
+        # Test case: middle case.
+        head = Node(0)
+        head.next = Node(1)
+        head.next.next = Node(2)
+        head.next.next.next = Node(3)
+        head.next.next.next.next = Node(4)
+        # Expect output: 0 4
+        self.assertEqual([0, 4], convert(remove_range(head, 1, 3)))
+
+        # Test case: taking out the front node
+        head = Node(0)
+        head.next = Node(1)
+        head.next.next = Node(2)
+        head.next.next.next = Node(3)
+        head.next.next.next.next = Node(4)
+        # Expect output: 2 3 4
+        self.assertEqual([2, 3, 4], convert(remove_range(head, 0, 1)))
+
+        # Test case: removing all the nodes
+        head = Node(0)
+        head.next = Node(1)
+        head.next.next = Node(2)
+        head.next.next.next = Node(3)
+        head.next.next.next.next = Node(4)
+        self.assertEqual([], convert(remove_range(head, 0, 7)))
+
+    def test_swap_in_pairs(self):
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(3)
+        head.next.next.next = Node(4)
+        # Expect output : 2 --> 1 --> 4 --> 3
+        self.assertEqual([2, 1, 4, 3], convert(swap_pairs(head)))
+
+    def test_rotate_right(self):
+        # Given 1->2->3->4->5->NULL
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(3)
+        head.next.next.next = Node(4)
+        head.next.next.next.next = Node(5)
+        # K = 2. Expect output: 4->5->1->2->3->NULL.
+        self.assertEqual([4, 5, 1, 2, 3], convert(rotate_right(head, 2)))
+
+    def test_is_cyclic(self):
+        # create linked list => A -> B -> C -> D -> E -> C
+        head = Node('A')
+        head.next = Node('B')
+        curr = head.next
+        cyclic_node = Node('C')
+        curr.next = cyclic_node
+        curr = curr.next
+        curr.next = Node('D')
+        curr = curr.next
+        curr.next = Node('E')
+        curr = curr.next
+        curr.next = cyclic_node
+        self.assertTrue(is_cyclic(head))
+
+        # create linked list 1 -> 2 -> 3 -> 4
+        head = Node(1)
+        curr = head
+        for i in range(2, 6):
+            curr.next = Node(i)
+            curr = curr.next
+        self.assertFalse(is_cyclic(head))
+
+if __name__ == "__main__":
+    unittest.main()
